[TreeSelect]废弃 searchInbox 属性，多选组件支持配置已选数据超长展示省略号
