+++ src/components/tree-select/demos/basic-tree-select.md
@@ -61,7 +61,6 @@ class TreeSelectDemo extends React.Component {
                 {/*<h5>单选搜索框在下拉框内</h5>*/}
                 <TreeSelect
                   searchable
-                  searchInBox={false}
                   allowClear
                   onSearch={this.handleSearch}
                   value={{
@@ -77,7 +76,6 @@ class TreeSelectDemo extends React.Component {
               {/*<div>*/}
               {/*  <h5>单选搜索框在下拉框外（新）</h5>*/}
               {/*  <TreeSelect*/}
-              {/*    searchInBox*/}
               {/*    searchable*/}
               {/*    allowClear*/}
               {/*    onSearch={this.handleSearch}*/}
