+++ src/components/tree-select/demos/dynamicLoad.md
@@ -100,7 +100,6 @@ class TreeSelectDemo extends React.Component {
                 <h5>单选</h5>
                 <TreeSelect
                   searchable
-                  searchInBox
                   allowClear
                   type="single"
                   containParentNode
@@ -146,7 +145,6 @@ class TreeSelectDemo extends React.Component {
                 <h5>多选</h5>
                 <TreeSelect
                   searchable
-                  searchInBox
                   allowClear
                   type="multiple"
                   containParentNode
